Add use rootless page (#93)
Tagging @cboettig since he suggested me to write something here

I have a solution for running the container rootless:

- https://github.com/rocker-org/rocker-versioned2/pull/636

There are many advantages on rootless containers, the main one being
security.

The main caveat with rootless containers is when we want to map
additional groups to the container (for instance when we have an
additional group that owns a "shared_data" directory we want to access).
In that case, we still need to learn quite a bit about id mapping. I've
done my best to explain how things work and to provide a step by step
guide in this pull request.

Hopefully this will eventually be simplified. It may be that I have
overlooked something

- https://github.com/containers/podman/issues/18333

I guess we can wait some days to see how the issue evolves. It may be
that I've missed something and my solution is overly complicated or that
some feature needs to land in podman to simplify additional group
management.

English is not my primary language. I would appreciate feedback or
change in wordings.

Besides I've been writing this for too long. I may need to take some
time to get some perspective and re-read it again, but I believe it is
worth a first read.

---------

Co-authored-by: eitsupi <50911393+eitsupi@users.noreply.github.com>